+++ src/main/java/com/bettrnation/model/BettingOutcomeResult.java
@@ -1,14 +1,10 @@
 package com.bettrnation.model;
 
-import java.math.BigDecimal;
-
 public record BettingOutcomeResult(
-        Integer bettingOutcomeId,
-        Integer bettingResultTypeId,
-        String bettingResultType,
-        Integer bettingOutcomeTypeId,
-        String bettingOutcomeType,
-        BigDecimal betValue,
-        BigDecimal actualValue
-) {
-}
+    Integer bettingOutcomeId,
+    Integer bettingResultTypeId,
+    String bettingResultType,
+    Integer bettingOutcomeTypeId,
+    String bettingOutcomeType,
+    Double betValue,
+    Double actualValue) {}
