+++ src/main/java/com/bettrnation/model/ConsensusOutcome.java
@@ -1,30 +1,27 @@
 package com.bettrnation.model;
 
-import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 public record ConsensusOutcome(
-        Integer numberOfSportsbooks,
-        Integer bettingOutcomeID,
-        Integer bettingMarketID,
-        Sportsbook sportsbook,
-        Integer bettingOutcomeTypeID,
-        String bettingOutcomeType,
-        Integer payoutAmerican,
-        BigDecimal payoutDecimal,
-        BigDecimal value,
-        String participant,
-        Boolean isAvailable,
-        Boolean isAlternate,
-        LocalDateTime created,
-        LocalDateTime updated,
-        LocalDateTime unlisted,
-        Integer teamID,
-        Integer playerID,
-        Integer globalTeamID,
-        String sportsbookUrl,
-        Boolean isInPlay,
-        String sportsbookMarketID,
-        String sportsbookOutcomeID
-) {
-}
+    Integer numberOfSportsbooks,
+    Integer bettingOutcomeID,
+    Integer bettingMarketID,
+    Sportsbook sportsbook,
+    Integer bettingOutcomeTypeID,
+    String bettingOutcomeType,
+    Integer payoutAmerican,
+    Double payoutDecimal,
+    Double value,
+    String participant,
+    Boolean isAvailable,
+    Boolean isAlternate,
+    LocalDateTime created,
+    LocalDateTime updated,
+    LocalDateTime unlisted,
+    Integer teamID,
+    Integer playerID,
+    Integer globalTeamID,
+    String sportsbookUrl,
+    Boolean isInPlay,
+    String sportsbookMarketID,
+    String sportsbookOutcomeID) {}
